gsgpu: Don't check for firmware file twice. Don't say GFX8.
Don't check for firmware file early in the driver loading code.
Also, some system may compress their kernel modules or have them
in a non-standard place so vendor's poor coding of the firmware
checking logic is improper.

Also don't say 'GFX8' in the error messages as that belongs to
AMD. The HW might be based on AMD designs but I imagine the vendor
might want to keep that from the public.